ThoughtsBackend/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Delorean==1.0.0
Faker==28.4.1
Flask==3.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
freezegun==1.5.1
greenlet==3.1.0
humanize==4.10.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
packaging==24.1
parse==1.20.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.0.4
